+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [WORKING ON Version 2015.01.01](https://github.com/texhex/testutil/releases/tag/v2015.01.01) 
+## [Version 2015.01.01](https://github.com/texhex/testutil/releases/tag/v2015.01.01) 
 
 **BREAKING CHANGES:**